Done
Now sending the booking total cost to the store and after that gets it from the payment component so that the booking total is sent to stripe as we wanted. I dont know if this is the best way to handle the problem but this was the solution i could come up with.